added more cout's to cpp and changed int to double in h

+++ src/pMotorCommand/MotorCommand.h
@@ -31,8 +31,8 @@ class MotorCommand : public CMOOSApp
    void CreateMotorHat(int);
    void CreateMotorObj(int);
 
-   int m_speed;
-   int m_command;
+   double m_speed;
+   double m_command;
    int m_I2C;
    int m_num;
    adafruit_motorhat m_mH;
